+++ src/main/java/com/capstone/smutaxi/config/SecurityConfig.java
@@ -65,8 +65,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/",
                         "/api/auth/join",
                         "/api/auth/login",
-                        "/api/auth/users/{email}",
-                        "/api/auth/users/{email}/password",
+                        "/api/update/users/{email}",
+                        "/api/update/users/{email}/password",
                         "/api/auth/join/email-verification",
                         "/api/auth/update/email-verification",
                         "/api/auth/check-duplicate/{email}",
